Chapter 3 - Insert and select with tables
I've created a new class for running selects as its it's already more complex than running the other statements.

The new table class is responsible keeping track of data and column definitions for a table.

To run tableless select queries, I'm creating a new table with a single empty row so it can run in the same way as other queries.