Chapter 7 - Propagate null
The simpler, but slightly repetitive, way of doing this would be to add null handling to each operator, or push the operations down to the nodes.

This has similarities to the way 'RETURNS NULL ON NULL INPUT' parameter to 'CREATE FUNCTION' works - if an argument is null, the function isn't executed.