Refactor - Make it easier to configure new functions and operators
Add the type information and the implementations of operators and functions closer together so they can keep in sync.

For operators that accept multiple types and numbers of input arguments, I've chosen to share an implementation for the different input. This makes looking up operators slightly easier as you don't need to know input types to find the operator.

Currently all operators and functions have a single return type that isn't dependent on their inputs. This isn't true for all functions (e.g COALESCE).