Refactor - Make constructors for null objects
It's handy to be able to create a new empty table and row metadata so they can be used when a real table or row metadata doesn't exist (such as when running tableless selects or when typing expressions that have no context. This avoids 'nil' checking elsewhere.

Note that the 'null' refers to the 'null object pattern' not the NULL in SQL.